requirements.txt
streamlit
langchain
google-generativeai
langchain-google-genai
langchain-community
python-dotenv
fitz
PyMuPDF
